hybris: egl: Move wlegl to common platform
