+++ hybris/Makefile.am
@@ -6,7 +6,7 @@ endif
 if HAS_ANDROID_5_0_0
 SUBDIRS += libsync
 endif
-SUBDIRS += egl glesv1 glesv2 ui sf input camera vibrator media opencl wifi hwc2
+SUBDIRS += platforms egl glesv1 glesv2 ui sf input camera vibrator media opencl wifi hwc2
 
 SUBDIRS += libnfc_nxp libnfc_ndef_nxp
