+++ hybris/egl/platforms/common/Makefile.am
@@ -2,38 +2,12 @@ lib_LTLIBRARIES = \
 	libhybris-eglplatformcommon.la
 
 libhybris_eglplatformcommon_la_SOURCES = \
-	native_handle.c \
-	nativewindowbase.cpp \
-	eglplatformcommon.cpp \
-	windowbuffer.cpp
+	eglplatformcommon.cpp
 
 pkgconfigdir = $(libdir)/pkgconfig
 pkgconfig_DATA = hybris-egl-platform.pc
 
-if WANT_WAYLAND
-
-libhybris_eglplatformcommon_la_SOURCES += \
-	server_wlegl.cpp \
-	server_wlegl_handle.cpp \
-	server_wlegl_buffer.cpp \
-	wayland-android-protocol.c
-
-BUILT_SOURCES = wayland-android-protocol.c \
-		wayland-android-client-protocol.h \
-		wayland-android-server-protocol.h
-
-%-protocol.c : %.xml
-	$(AM_V_GEN)$(WAYLAND_SCANNER) code < $< > $@
-
-%-server-protocol.h : %.xml
-	$(AM_V_GEN)$(WAYLAND_SCANNER) server-header < $< > $@
-
-%-client-protocol.h : %.xml
-	$(AM_V_GEN)$(WAYLAND_SCANNER) client-header < $< > $@
-
-endif
-
-libhybris_eglplatformcommon_la_CFLAGS = -I$(top_srcdir)/include $(ANDROID_HEADERS_CFLAGS) -I$(top_srcdir)/egl -I$(top_srcdir)/common/
+libhybris_eglplatformcommon_la_CFLAGS = -I$(top_srcdir)/include $(ANDROID_HEADERS_CFLAGS) -I$(top_srcdir)/egl -I$(top_srcdir)/common/ -I$(top_srcdir)/platforms/common/
 if WANT_WAYLAND
 libhybris_eglplatformcommon_la_CFLAGS += $(WAYLAND_CLIENT_CFLAGS) $(WAYLAND_SERVER_CFLAGS) $(ANDROID_HEADERS_CFLAGS)
 endif
@@ -49,7 +23,7 @@ libhybris_eglplatformcommon_la_CFLAGS += -ggdb -O0
 endif
 
 
-libhybris_eglplatformcommon_la_CXXFLAGS = -I$(top_srcdir)/include -I$(top_srcdir)/egl $(ANDROID_HEADERS_CFLAGS) -I$(top_srcdir)/common/
+libhybris_eglplatformcommon_la_CXXFLAGS = -I$(top_srcdir)/include -I$(top_srcdir)/egl $(ANDROID_HEADERS_CFLAGS) -I$(top_srcdir)/common/ -I$(top_srcdir)/platforms/common/
 if WANT_MESA
 libhybris_eglplatformcommon_la_CXXFLAGS += -DLIBHYBRIS_WANTS_MESA_X11_HEADERS
 endif
@@ -66,6 +40,7 @@ libhybris_eglplatformcommon_la_CXXFLAGS += -ggdb -O0
 endif
 libhybris_eglplatformcommon_la_LDFLAGS = \
 	$(top_builddir)/common/libhybris-common.la \
+	$(top_builddir)/platforms/common/libhybris-platformcommon.la \
 	-version-info "1":"0":"0"
 
 if HAS_ANDROID_4_2_0
@@ -79,9 +54,8 @@ endif
 
 eglplatformcommondir = $(includedir)/hybris/eglplatformcommon
 eglplatformcommon_HEADERS = \
-	support.h \
-	nativewindowbase.h \
 	eglplatformcommon.h \
+	eglnativewindowbase.h \
 	hybris_nativebufferext.h
 
 if WANT_WAYLAND
