+++ hybris/egl/platforms/fbdev/Makefile.am
@@ -7,6 +7,7 @@ eglplatform_fbdev_la_SOURCES = \
 eglplatform_fbdev_la_CXXFLAGS = \
 	-I$(top_srcdir)/include \
 	-I$(top_srcdir)/common \
+	-I$(top_srcdir)/platforms/common \
 	-I$(top_srcdir)/egl \
 	-I$(top_srcdir)/egl/platforms/common \
 	$(ANDROID_HEADERS_CFLAGS)
