+++ hybris/egl/platforms/hwcomposer/Makefile.am
@@ -14,6 +14,7 @@ libhybris_hwcomposerwindow_la_SOURCES = \
 libhybris_hwcomposerwindow_la_CXXFLAGS = \
 	-I$(top_srcdir)/include \
 	-I$(top_srcdir)/common \
+	-I$(top_srcdir)/platforms/common \
 	-I$(top_srcdir)/egl \
 	-I$(top_srcdir)/egl/platforms/common \
 	$(ANDROID_HEADERS_CFLAGS)
@@ -44,6 +45,7 @@ eglplatform_hwcomposer_la_SOURCES = \
 eglplatform_hwcomposer_la_CXXFLAGS = \
 	-I$(top_srcdir)/include \
 	-I$(top_srcdir)/common \
+	-I$(top_srcdir)/platforms/common \
 	-I$(top_srcdir)/egl \
 	-I$(top_srcdir)/egl/platforms/common \
 	$(ANDROID_HEADERS_CFLAGS)
