+++ hybris/egl/platforms/wayland/Makefile.am
@@ -3,6 +3,7 @@ pkglib_LTLIBRARIES = eglplatform_wayland.la
 eglplatform_wayland_la_SOURCES = eglplatform_wayland.cpp wayland_window.cpp
 eglplatform_wayland_la_CXXFLAGS = \
 	-I$(top_srcdir)/common \
+	-I$(top_srcdir)/platforms/common \
 	-I$(top_srcdir)/include \
 	-I$(top_srcdir)/egl \
 	-I$(top_srcdir)/egl/platforms/common \
