+++ hybris/platforms/Makefile.am
@@ -1,0 +1,1 @@
+SUBDIRS = common
