+++ hybris/platforms/common/native_handle.c
