+++ hybris/platforms/common/nativewindowbase.cpp
@@ -1,3 +1,18 @@
+/*
+ * Copyright (c) 2013-2022 Jolla Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 
 #ifndef ANDROID_BUILD
 #include <android-config.h>
@@ -340,7 +355,7 @@ int BaseNativeWindow::_query(const struct ANativeWindow* window, int what, int* 
 			return NO_ERROR;
 #endif
 	}
-	TRACE("EGL error: unkown window attribute! %i", what);
+	TRACE("NativeWindow error: unknown window attribute! %i", what);
 	*value = 0;
 	return BAD_VALUE;
 }
