+++ hybris/platforms/common/support.h
