+++ hybris/platforms/common/wayland-android.xml
