+++ hybris/platforms/common/windowbuffer.cpp
