+++ hybris/platforms/common/windowbuffer.h
